Remove legacy pylons code
