+++ MANIFEST.in
@@ -1,3 +1,4 @@
 recursive-include ckanext/spatial/templates                                           *
 recursive-include ckanext/spatial/public                                              *
+recursive-include ckanext/spatial/assets                                              *
 recursive-include ckanext/spatial/validation/xml                                      *
