+++ ckanext/spatial/assets/css/common_map.css
