+++ ckanext/spatial/assets/css/dataset_map.css
