+++ ckanext/spatial/assets/css/spatial_query.css
