+++ ckanext/spatial/assets/js/common_map.js
