+++ ckanext/spatial/assets/js/dataset_map.js
