+++ ckanext/spatial/assets/js/spatial_query.js
