+++ ckanext/spatial/assets/webassets.yml
