--- ckanext/spatial/public/resource.config
@@ -1,25 +1,0 @@
-[depends]
-
-main = base/main
-
-[groups]
-
-dataset_map =
-
-    js/vendor/leaflet/1.9.3/leaflet.js
-    js/common_map.js
-    js/dataset_map.js
-
-    js/vendor/leaflet/1.9.3/leaflet.css
-    css/dataset_map.css
-
-spatial_query =
-
-    js/vendor/leaflet/1.9.3/leaflet.js
-    js/vendor/leaflet.draw/0.4.14/leaflet.draw.js
-    js/common_map.js
-    js/spatial_query.js
-
-    js/vendor/leaflet/1.9.3/leaflet.css
-    js/vendor/leaflet.draw/0.4.14/leaflet.draw.css
-    css/spatial_query.css
