--- ckanext/spatial/templates/spatial/snippets/dataset_map_asset.html
@@ -1,2 +1,0 @@
-{% asset 'ckanext-spatial/dataset_map_js' %}
-{% asset 'ckanext-spatial/dataset_map_css' %}
