--- ckanext/spatial/templates/spatial/snippets/dataset_map_resource.html
@@ -1,1 +1,0 @@
-{% resource 'ckanext-spatial/dataset_map' %}
