+++ ckanext/spatial/templates/spatial/snippets/spatial_query.html
@@ -24,5 +24,5 @@ e.g.
     </div>
 </section>
 
-{% set type = 'asset' if h.ckan_version().split('.')[1] | int >= 9 else 'resource' %}
-{% include 'spatial/snippets/spatial_query_' ~ type ~ '.html' %}
+{% asset 'ckanext-spatial/spatial_query_js' %}
+{% asset 'ckanext-spatial/spatial_query_css' %}
