--- ckanext/spatial/templates/spatial/snippets/spatial_query_asset.html
@@ -1,2 +1,0 @@
-{% asset 'ckanext-spatial/spatial_query_js' %}
-{% asset 'ckanext-spatial/spatial_query_css' %}
