--- ckanext/spatial/templates/spatial/snippets/spatial_query_resource.html
@@ -1,1 +1,0 @@
-{% resource 'ckanext-spatial/spatial_query' %}
